backend/requirements.txt
asgiref==3.7.2
Django==4.1.13
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
pillow==10.2.0
pymongo==3.12.3
python-decouple==3.8
pytz==2023.4
sqlparse==0.2.4
tzdata==2023.4
